# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/course/{courseId}/lessons": {
            "get": {
                "summary": "Retrieves all lessons for a specific course",
                "operationId": "2bcb5a5a16ddedf503b59006d773a1ce",
                "parameters": [
                    {
                        "name": "courseId",
                        "in": "path",
                        "description": "The ID of the course to retrieve lessons from",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval of lessons",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID of the lesson",
                                                "type": "integer"
                                            },
                                            "course_id": {
                                                "description": "ID of the course the lesson belongs to",
                                                "type": "integer"
                                            },
                                            "title": {
                                                "description": "Title of the lesson",
                                                "type": "string"
                                            },
                                            "content": {
                                                "description": "Content of the lesson",
                                                "type": "string"
                                            },
                                            "created_at": {
                                                "description": "Creation date of the lesson",
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "description": "Last update date of the lesson",
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No lessons found for the course",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Error message indicating no lessons found",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Error message detailing an internal error",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/course/lesson/create": {
            "post": {
                "summary": "Creates a new lesson for a specified course",
                "operationId": "beedfc949d532e763e9511795ec8ee9c",
                "requestBody": {
                    "description": "Data and video needed to create a new lesson",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "course_id",
                                    "title",
                                    "video_url",
                                    "content",
                                    "duration"
                                ],
                                "properties": {
                                    "course_id": {
                                        "description": "ID of the course this lesson belongs to",
                                        "type": "integer"
                                    },
                                    "title": {
                                        "description": "Title of the lesson",
                                        "type": "string"
                                    },
                                    "video_url": {
                                        "description": "Video file for the lesson",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "content": {
                                        "description": "Content of the lesson",
                                        "type": "string"
                                    },
                                    "duration": {
                                        "description": "Duration of the video in seconds",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lesson created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID of the newly created lesson",
                                            "type": "integer"
                                        },
                                        "video_url": {
                                            "description": "URL of the lesson video",
                                            "type": "string"
                                        },
                                        "title": {
                                            "description": "Title of the lesson",
                                            "type": "string"
                                        },
                                        "content": {
                                            "description": "Content of the lesson",
                                            "type": "string"
                                        },
                                        "duration": {
                                            "description": "Duration of the lesson in HH:MM:SS format",
                                            "type": "string"
                                        },
                                        "course_id": {
                                            "description": "ID of the course the lesson belongs to",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Error message detailing the validation error",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Error message indicating an internal error occurred",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/course/{courseId}/lesson": {
            "get": {
                "summary": "Retrieves all lessons for a specific course",
                "operationId": "2005da70e38c4d4dd044fc332e3bb7d7",
                "parameters": [
                    {
                        "name": "courseId",
                        "in": "path",
                        "description": "The ID of the course to retrieve lessons from",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval of lessons",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID of the lesson",
                                                "type": "integer"
                                            },
                                            "course_id": {
                                                "description": "ID of the course the lesson belongs to",
                                                "type": "integer"
                                            },
                                            "title": {
                                                "description": "Title of the lesson",
                                                "type": "string"
                                            },
                                            "content": {
                                                "description": "Content of the lesson",
                                                "type": "string"
                                            },
                                            "created_at": {
                                                "description": "Creation date of the lesson",
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "description": "Last update date of the lesson",
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No lessons found for the course",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Error message indicating no lessons found",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Error message detailing an internal error",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}